Working vite + swc
