+++ website-swc/src/components/CopyButton.jsx
