+++ website-swc/src/tabs/advanced/EncryptAccount.jsx
