+++ website-swc/src/tabs/develop/Split.jsx
@@ -1,37 +1,36 @@
 import React, { useState, useEffect } from "react";
 import {Button, Card, Col, Divider, Form, Input, Row, Result, Spin, Switch} from "antd";
 import axios from "axios";
+import AleoWorker from '../../workers/worker.js?worker'
 
-export const Join = () => {
-    const [joinFeeRecord, setJoinFeeRecord] = useState(null);
-    const [recordOne, setRecordOne] = useState(null);
-    const [recordTwo, setRecordTwo] = useState(null);
-    const [joinUrl, setJoinUrl] = useState("https://vm.aleo.org/api");
-    const [joinFee, setJoinFee] = useState("1.0");
+export const Split = () => {
+    const [amountRecord, setAmountRecord] = useState(null);
+    const [splitUrl, setSplitUrl] = useState("https://vm.aleo.org/api");
+    const [splitAmount, setSplitAmount] = useState("1.0");
     const [loading, setLoading] = useState(false);
     const [privateKey, setPrivateKey] = useState(null);
-    const [joinError, setJoinError] = useState(null);
+    const [splitError, setSplitError] = useState(null);
     const [status, setStatus] = useState("");
     const [transactionID, setTransactionID] = useState(null);
     const [worker, setWorker] = useState(null);
 
     function spawnWorker() {
-        let worker = new Worker("./worker.js");
+        let worker = new AleoWorker();
         worker.addEventListener("message", ev => {
-            if (ev.data.type == 'JOIN_TRANSACTION_COMPLETED') {
-                axios.post(peerUrl() + "/testnet3/transaction/broadcast", ev.data.joinTransaction, {
+            if (ev.data.type == 'SPLIT_TRANSACTION_COMPLETED') {
+                axios.post(peerUrl() + "/testnet3/transaction/broadcast", ev.data.splitTransaction, {
                     headers: {
                         'Content-Type': 'application/json',
                     }
                 }).then(
                     (response) => {
                         setLoading(false);
-                        setJoinError(null);
+                        setSplitError(null);
                         setTransactionID(response.data);
                     }
                 )
             } else if (ev.data.type == 'ERROR') {
-                setJoinError(ev.data.errorMessage);
+                setSplitError(ev.data.errorMessage);
                 setLoading(false);
                 setTransactionID(null);
             }
@@ -49,28 +48,26 @@ export const Join = () => {
         }
     }, []);
 
-    const join = async (event) => {
+    const split = async (event) => {
         setLoading(true)
         setTransactionID(null);
-        setJoinError(null);
+        setSplitError(null);
 
-        const feeAmount = parseFloat(feeString());
-        if (isNaN(feeAmount)) {
-            setJoinError("Fee is not a valid number");
+        const amount = parseFloat(amountString());
+        if (isNaN(amount)) {
+            setSplitError("Amount is not a valid number");
             setLoading(false);
             return;
-        } else if (feeAmount <= 0) {
-            setJoinError("Fee must be greater than 0");
+        } else if (amount <= 0) {
+            setSplitError("Amount must be greater than 0");
             setLoading(false);
             return;
         }
 
         await postMessagePromise(worker, {
-            type: 'ALEO_JOIN',
-            recordOne: recordOneString(),
-            recordTwo: recordTwoString(),
-            fee: feeAmount,
-            feeRecord: feeRecordString(),
+            type: 'ALEO_SPLIT',
+            splitAmount: amount,
+            record: amountRecordString(),
             privateKey: privateKeyString(),
             url: peerUrl(),
         });
@@ -82,7 +79,7 @@ export const Join = () => {
                 resolve(event.data);
             };
             worker.onerror = error => {
-                setJoinError(error);
+                setSplitError(error);
                 setLoading(false);
                 setTransactionID(null);
                 reject(error);
@@ -93,45 +90,27 @@ export const Join = () => {
 
     const onUrlChange = (event) => {
         if (event.target.value !== null) {
-            setJoinUrl(event.target.value);
+            setSplitUrl(event.target.value);
         }
-        return joinUrl;
+        return splitUrl;
     }
 
-    const onJoinFeeChange = (event) => {
+    const onAmountChange = (event) => {
         if (event.target.value !== null) {
-            setJoinFee(event.target.value);
+            setSplitAmount(event.target.value);
         }
         setTransactionID(null);
-        setJoinError(null);
-        return joinFee;
+        setSplitError(null);
+        return splitAmount;
     }
 
-    const onRecordOneChange = (event) => {
+    const onAmountRecordChange = (event) => {
         if (event.target.value !== null) {
-            setRecordOne(event.target.value);
+            setAmountRecord(event.target.value);
         }
         setTransactionID(null);
-        setJoinError(null);
-        return recordOne;
-    }
-
-    const onRecordTwoChange = (event) => {
-        if (event.target.value !== null) {
-            setRecordTwo(event.target.value);
-        }
-        setTransactionID(null);
-        setJoinError(null);
-        return recordTwo;
-    }
-
-    const onJoinFeeRecordChange = (event) => {
-        if (event.target.value !== null) {
-            setJoinFeeRecord(event.target.value);
-        }
-        setTransactionID(null);
-        setJoinError(null);
-        return joinFeeRecord;
+        setSplitError(null);
+        return amountRecord;
     }
 
     const onPrivateKeyChange = (event) => {
@@ -139,70 +118,45 @@ export const Join = () => {
             setPrivateKey(event.target.value);
         }
         setTransactionID(null);
-        setJoinError(null);
+        setSplitError(null);
         return privateKey;
     }
 
     const layout = { labelCol: { span: 3 }, wrapperCol: { span: 21 } };
+    const amountString = () => splitAmount !== null ? splitAmount : "";
     const privateKeyString = () => privateKey !== null ? privateKey : "";
-    const feeRecordString = () => joinFeeRecord !== null ? joinFeeRecord : "";
-    const recordOneString = () => recordOne !== null ? recordOne : "";
-    const recordTwoString = () => recordTwo !== null ? recordTwo : "";
+    const amountRecordString = () => amountRecord !== null ? amountRecord : "";
     const transactionIDString = () => transactionID !== null ? transactionID : "";
-    const joinErrorString = () => joinError !== null ? joinError : "";
-    const feeString = () => joinFee !== null ? joinFee : "";
-    const peerUrl = () => joinUrl !== null ? joinUrl : "";
+    const splitErrorString = () => splitError !== null ? splitError : "";
+    const peerUrl = () => splitUrl !== null ? splitUrl : "";
 
-    return <Card title="Join Records"
+
+    return <Card title="Split Record"
                  style={{width: "100%", borderRadius: "20px"}}
                  bordered={false}>
         <Form {...layout}>
-            <Form.Item label="Record One"
+            <Form.Item label="Split Amount"
                        colon={false}
                        validateStatus={status}
             >
-                <Input.TextArea name="Record One"
-                                size="small"
-                                placeholder="First Record to Join"
+                <Input.TextArea name="split amount"
+                                size="large"
+                                placeholder="Amount to split record into"
                                 allowClear
-                                onChange={onRecordOneChange}
-                                value={recordOneString()}
+                                onChange={onAmountChange}
+                                value={amountString()}
                                 style={{borderRadius: '20px'}}/>
             </Form.Item>
-            <Form.Item label="Record Two"
+            <Form.Item label="Record"
                        colon={false}
                        validateStatus={status}
             >
-                <Input.TextArea name="Record Two"
+                <Input.TextArea name="Record"
                                 size="small"
-                                placeholder="Second Record to Join"
+                                placeholder="Record to split"
                                 allowClear
-                                onChange={onRecordTwoChange}
-                                value={recordTwoString()}
-                                style={{borderRadius: '20px'}}/>
-            </Form.Item>
-            <Form.Item label="Fee"
-                       colon={false}
-                       validateStatus={status}
-            >
-                <Input.TextArea name="Fee"
-                                size="small"
-                                placeholder="Fee"
-                                allowClear
-                                onChange={onJoinFeeChange}
-                                value={feeString()}
-                                style={{borderRadius: '20px'}}/>
-            </Form.Item>
-            <Form.Item label="Fee Record"
-                       colon={false}
-                       validateStatus={status}
-            >
-                <Input.TextArea name="Fee Record"
-                                size="small"
-                                placeholder="Record used to pay join fee"
-                                allowClear
-                                onChange={onJoinFeeRecordChange}
-                                value={feeRecordString()}
+                                onChange={onAmountRecordChange}
+                                value={amountRecordString()}
                                 style={{borderRadius: '20px'}}/>
             </Form.Item>
             <Form.Item label="Private Key"
@@ -231,30 +185,30 @@ export const Join = () => {
             </Form.Item>
             <Row justify="center">
                 <Col justify="center">
-                    <Button type="primary" shape="round" size="middle" onClick={join}
-                    >Join</Button>
+                    <Button type="primary" shape="round" size="middle" onClick={split}
+                    >Split</Button>
                 </Col>
             </Row>
         </Form>
         <Row justify="center" gutter={[16, 32]} style={{ marginTop: '48px' }}>
             {
                 (loading === true) &&
-                <Spin tip="Creating Join..." size="large"/>
+                <Spin tip="Creating Split..." size="large"/>
             }
             {
                 (transactionID !== null) &&
                 <Result
                     status="success"
-                    title="Join Successful!"
+                    title="Split Successful!"
                     subTitle={"Transaction ID: " + transactionIDString()}
                 />
             }
             {
-                (joinError !== null) &&
+                (splitError !== null) &&
                 <Result
                     status="error"
-                    title="Join Error"
-                    subTitle={"Error: " + joinErrorString()}
+                    title="Split Error"
+                    subTitle={"Error: " + splitErrorString()}
                 />
             }
         </Row>
