+++ website-swc/src/tabs/record/DecryptRecord.jsx
